+++ src/navigation/AppNavigator.tsx
@@ -13,22 +13,22 @@ import {
   MainTabParamList,
 } from './types';
 import LandingScreen from '../screens/LandingScreen';
-import LoginScreen from '../screens/LoginScreen';
+import LoginScreen from '../screens/auth/LoginScreen';
 import HomeScreen from '../screens/HomeScreen';
-import SignupScreen from '../screens/SignupScreen';
-import UserDetailScreen from '../screens/UserDetailScreen';
+import SignupScreen from '../screens/auth/SignupScreen';
+import UserDetailScreen from '../screens/auth/UserDetailScreen';
 import ReservationScreen from '../screens/ReservationScreen';
 import NewPaxiRoomScreen from '../screens/NewPaxiRoomScreen';
 import PaxiRoomListScreen from '../screens/PaxiRoomListScreen';
-import PlaceReservationScreen from '../screens/PlaceReservationScreen';
-import PlaceDetailReservationScreen from '../screens/PlaceDetailReservationScreen';
-import PlaceReservationApplyScreen from '../screens/PlaceReservationApplyScreen';
+import PlaceReservationScreen from '../screens/place-reservation/PlaceReservationScreen';
+import PlaceDetailReservationScreen from '../screens/place-reservation/PlaceDetailReservationScreen';
+import PlaceReservationApplyScreen from '../screens/place-reservation/PlaceReservationApplyScreen';
 import WhitebookScreen from '../screens/WhitebookScreen';
 import BenefitsScreen from '../screens/BenefitsScreen';
-import ClubScreen from '../screens/ClubScreen';
-import ClubDetailScreen from '../screens/ClubDetailScreen';
-import AssociationScreen from '../screens/AssociationScreen';
-import AssociationDetailScreen from '../screens/AssociationDetailScreen';
+import ClubScreen from '../screens/club/ClubScreen';
+import ClubDetailScreen from '../screens/club/ClubDetailScreen';
+import AssociationScreen from '../screens/association/AssociationScreen';
+import AssociationDetailScreen from '../screens/association/AssociationDetailScreen';
 import CampusShuttleScreen from '../screens/CampusShuttle';
 
 const Stack = createNativeStackNavigator<RootStackParamList>();
