+++ src/screens/club/ClubDetailScreen.tsx
@@ -13,24 +13,24 @@ import {
 import {SafeAreaView} from 'react-native-safe-area-context';
 import {NativeStackNavigationProp} from '@react-navigation/native-stack';
 import {RouteProp} from '@react-navigation/native';
-import {RootStackParamList} from '../navigation/types';
-import api from '../utils/api';
+import {RootStackParamList} from '../../navigation/types';
+import {ClubType} from '../../types/club';
+import api from '../../utils/api';
 
-type AssociationDetailScreenProps = {
-  navigation: NativeStackNavigationProp<
-    RootStackParamList,
-    'AssociationDetail'
-  >;
-  route: RouteProp<RootStackParamList, 'AssociationDetail'>;
+type ClubDetailScreenProps = {
+  navigation: NativeStackNavigationProp<RootStackParamList, 'ClubDetail'>;
+  route: RouteProp<RootStackParamList, 'ClubDetail'>;
 };
 
-interface AssociationItem {
+interface ClubItem {
   uuid: string;
   name: string;
+  short_desc: string;
   content: string;
   location: string;
   representative: string;
   contact: string;
+  clubType: ClubType;
   image_url: string;
   views: number;
   homepage_url: string;
@@ -39,33 +39,30 @@ interface AssociationItem {
   youtube_url: string;
 }
 
-const AssociationDetailScreen: React.FC<AssociationDetailScreenProps> = ({
+const ClubDetailScreen: React.FC<ClubDetailScreenProps> = ({
   navigation,
   route,
 }) => {
   const isDarkMode = useColorScheme() === 'dark';
-  const [associationData, setAssociationData] =
-    useState<AssociationItem | null>(null);
+  const [clubData, setClubData] = useState<ClubItem | null>(null);
   const [isLoading, setIsLoading] = useState(true);
 
-  const {associationId, associationName} = route.params;
+  const {clubId, clubName} = route.params;
 
   useEffect(() => {
-    const fetchAssociationDetail = async () => {
+    const fetchClubDetail = async () => {
       try {
-        const response = await api.get<AssociationItem>(
-          `/introduce/association/${associationId}`,
-        );
-        setAssociationData(response.data);
+        const response = await api.get<ClubItem>(`/introduce/club/${clubId}`);
+        setClubData(response.data);
       } catch (error) {
-        console.error('자치단체 상세 정보 로드 오류:', error);
+        console.error('동아리 상세 정보 로드 오류:', error);
       } finally {
         setIsLoading(false);
       }
     };
 
-    fetchAssociationDetail();
-  }, [associationId]);
+    fetchClubDetail();
+  }, [clubId]);
 
   const backgroundStyle = {
     backgroundColor: isDarkMode ? '#121212' : '#fff',
@@ -109,25 +106,30 @@ const AssociationDetailScreen: React.FC<AssociationDetailScreenProps> = ({
           onPress={() => navigation.goBack()}>
           <Text style={[styles.backButtonText, {color: textColor}]}>뒤로</Text>
         </TouchableOpacity>
-        <Text style={[styles.headerTitle, {color: textColor}]}>
-          {associationName}
-        </Text>
+        <Text style={[styles.headerTitle, {color: textColor}]}>{clubName}</Text>
         <View style={styles.placeholderButton} />
       </View>
 
       <ScrollView style={styles.container}>
-        {associationData && (
+        {clubData && (
           <>
             <View style={styles.imageContainer}>
               <Image
-                source={{uri: associationData.image_url}}
-                style={styles.associationImage}
+                source={{uri: clubData.image_url}}
+                style={styles.clubImage}
                 resizeMode="contain"
               />
             </View>
             <View style={styles.contentContainer}>
-              <Text style={[styles.associationName, {color: textColor}]}>
-                {associationData.name}
+              <Text style={[styles.clubName, {color: textColor}]}>
+                {clubData.name}
+              </Text>
+              <Text
+                style={[
+                  styles.description,
+                  {color: isDarkMode ? '#888' : '#666'},
+                ]}>
+                {clubData.short_desc}
               </Text>
 
               <View style={[styles.infoSection, {borderColor}]}>
@@ -136,7 +138,7 @@ const AssociationDetailScreen: React.FC<AssociationDetailScreenProps> = ({
                     styles.content,
                     {color: isDarkMode ? '#888' : '#666'},
                   ]}>
-                  {associationData.content}
+                  {clubData.content}
                 </Text>
               </View>
 
@@ -150,7 +152,7 @@ const AssociationDetailScreen: React.FC<AssociationDetailScreenProps> = ({
                       styles.infoValue,
                       {color: isDarkMode ? '#888' : '#666'},
                     ]}>
-                    {associationData.location}
+                    {clubData.location}
                   </Text>
                 </View>
                 <View style={styles.infoRow}>
@@ -162,7 +164,7 @@ const AssociationDetailScreen: React.FC<AssociationDetailScreenProps> = ({
                       styles.infoValue,
                       {color: isDarkMode ? '#888' : '#666'},
                     ]}>
-                    {associationData.representative}
+                    {clubData.representative}
                   </Text>
                 </View>
                 <View style={styles.infoRow}>
@@ -174,66 +176,58 @@ const AssociationDetailScreen: React.FC<AssociationDetailScreenProps> = ({
                       styles.infoValue,
                       {color: isDarkMode ? '#888' : '#666'},
                     ]}>
-                    {associationData.contact}
+                    {clubData.contact}
                   </Text>
                 </View>
               </View>
 
               <View style={[styles.infoSection, {borderColor}]}>
                 <View style={styles.socialLinks}>
-                  {associationData.homepage_url != 'null' &&
-                    associationData.homepage_url != '' &&
-                    associationData.homepage_url && (
+                  {clubData.homepage_url != 'null' &&
+                    clubData.homepage_url != '' &&
+                    clubData.homepage_url && (
                       <TouchableOpacity
                         style={[
                           styles.socialButton,
                           {backgroundColor: '#000000'},
                         ]}
-                        onPress={() =>
-                          handleLinkPress(associationData.homepage_url)
-                        }>
+                        onPress={() => handleLinkPress(clubData.homepage_url)}>
                         <Text style={styles.socialText}>홈페이지</Text>
                       </TouchableOpacity>
                     )}
-                  {associationData.facebook_url != 'null' &&
-                    associationData.facebook_url != '' &&
-                    associationData.facebook_url && (
+                  {clubData.facebook_url != 'null' &&
+                    clubData.facebook_url != '' &&
+                    clubData.facebook_url && (
                       <TouchableOpacity
                         style={[
                           styles.socialButton,
                           {backgroundColor: '#1877F2'},
                         ]}
-                        onPress={() =>
-                          handleLinkPress(associationData.facebook_url)
-                        }>
+                        onPress={() => handleLinkPress(clubData.facebook_url)}>
                         <Text style={styles.socialText}>Facebook</Text>
                       </TouchableOpacity>
                     )}
-                  {associationData.instagram_url != 'null' &&
-                    associationData.instagram_url != '' &&
-                    associationData.instagram_url && (
+                  {clubData.instagram_url != 'null' &&
+                    clubData.instagram_url != '' &&
+                    clubData.instagram_url && (
                       <TouchableOpacity
                         style={[
                           styles.socialButton,
                           {backgroundColor: '#E4405F'},
                         ]}
-                        onPress={() =>
-                          handleLinkPress(associationData.instagram_url)
-                        }>
+                        onPress={() => handleLinkPress(clubData.instagram_url)}>
                         <Text style={styles.socialText}>Instagram</Text>
                       </TouchableOpacity>
                     )}
-                  {associationData.youtube_url != 'null' &&
-                    associationData.youtube_url != '' &&
-                    associationData.youtube_url && (
+                  {clubData.youtube_url != 'null' &&
+                    clubData.youtube_url != '' &&
+                    clubData.youtube_url && (
                       <TouchableOpacity
                         style={[
                           styles.socialButton,
                           {backgroundColor: '#FF0000'},
                         ]}
-                        onPress={() =>
-                          handleLinkPress(associationData.youtube_url)
-                        }>
+                        onPress={() => handleLinkPress(clubData.youtube_url)}>
                         <Text style={styles.socialText}>YouTube</Text>
                       </TouchableOpacity>
                     )}
@@ -291,7 +285,7 @@ const styles = StyleSheet.create({
     overflow: 'hidden',
     marginBottom: 24,
   },
-  associationImage: {
+  clubImage: {
     padding: 10,
     width: '100%',
     height: '100%',
@@ -299,14 +293,24 @@ const styles = StyleSheet.create({
   contentContainer: {
     padding: 24,
   },
-  associationName: {
+  clubName: {
     fontSize: 24,
     fontWeight: 'bold',
     marginBottom: 12,
   },
+  description: {
+    fontSize: 16,
+    lineHeight: 24,
+    marginBottom: 32,
+  },
   infoSection: {
     borderTopWidth: 1,
     paddingVertical: 24,
+  },
+  sectionTitle: {
+    fontSize: 18,
+    fontWeight: '600',
+    marginBottom: 12,
   },
   content: {
     fontSize: 16,
@@ -346,6 +350,10 @@ const styles = StyleSheet.create({
     fontWeight: '600',
     textAlign: 'center',
   },
+  socialIcon: {
+    width: 100,
+    height: 28,
+  },
 });
 
-export default AssociationDetailScreen;
+export default ClubDetailScreen;
