+++ src/screens/club/ClubScreen.tsx
@@ -11,20 +11,43 @@ import {
 } from 'react-native';
 import {SafeAreaView} from 'react-native-safe-area-context';
 import {NativeStackNavigationProp} from '@react-navigation/native-stack';
-import {RootStackParamList} from '../navigation/types';
-import api from '../utils/api';
+import {RootStackParamList} from '../../navigation/types';
+import {ClubType} from '../../types/club';
+import api from '../../utils/api';
 
-type AssociationScreenProps = {
-  navigation: NativeStackNavigationProp<RootStackParamList, 'Association'>;
+type ClubScreenProps = {
+  navigation: NativeStackNavigationProp<RootStackParamList, 'Club'>;
 };
 
-interface AssociationItem {
+const clubCategories = [
+  ClubType.performance1,
+  ClubType.performance2,
+  ClubType.sports,
+  ClubType.hobbyAndExhibition,
+  ClubType.study,
+  ClubType.societyAndReligion,
+];
+
+const clubCategoryMap = {
+  performance1: '공연1',
+  performance2: '공연2',
+  sports: '체육',
+  hobbyAndExhibition: '취미전시',
+  study: '학술',
+  societyAndReligion: '사회종교',
+};
+
+const sortTypes = ['가나다순', '조회순'];
+
+interface ClubItem {
   uuid: string;
   name: string;
+  short_desc: string;
   content: string;
   location: string;
   representative: string;
   contact: string;
+  clubType: ClubType;
   image_url: string;
   views: number;
   homepage_url: string;
@@ -33,13 +56,15 @@ interface AssociationItem {
   youtube_url: string;
 }
 
-const sortTypes = ['가나다순', '조회순'];
-
-const AssociationScreen: React.FC<AssociationScreenProps> = ({navigation}) => {
+const ClubScreen: React.FC<ClubScreenProps> = ({navigation}) => {
   const isDarkMode = useColorScheme() === 'dark';
+  const [selectedCategory, setSelectedCategory] = useState<ClubType>(
+    ClubType.performance1,
+  );
+  const [textWidths, setTextWidths] = useState<{[key: string]: number}>({});
+  const [clubs, setClubs] = useState<ClubItem[]>([]);
   const [sortType, setSortType] = useState('가나다순');
   const [isLoading, setIsLoading] = useState(true);
-  const [associations, setAssociations] = useState<AssociationItem[]>([]);
 
   const backgroundStyle = {
     backgroundColor: isDarkMode ? '#121212' : '#fff',
@@ -49,30 +74,32 @@ const AssociationScreen: React.FC<AssociationScreenProps> = ({navigation}) => {
   const textColor = isDarkMode ? '#FFFFFF' : '#000000';
   const borderColor = isDarkMode ? '#2C2C2C' : '#E5E7EB';
 
-  const getSortedAssociations = (items: AssociationItem[]) => {
-    if (sortType === '가나다순') {
-      return [...items].sort((a, b) => a.name.localeCompare(b.name));
-    } else {
-      return [...items].sort((a, b) => b.views - a.views);
-    }
-  };
-
   useEffect(() => {
-    const fetchAssociations = async () => {
+    const fetchClubs = async () => {
       try {
-        const response = await api.get<AssociationItem[]>(
-          '/introduce/association',
-        );
-        setAssociations(response.data);
+        const response = await api.get<ClubItem[]>('/introduce/club');
+        setClubs(response.data);
       } catch (error) {
-        console.error('자치단체 데이터 로드 오류:', error);
+        console.error('동아리 데이터 로드 오류:', error);
       } finally {
         setIsLoading(false);
       }
     };
 
-    fetchAssociations();
+    fetchClubs();
   }, []);
+
+  const getSortedClubs = (clubs: ClubItem[]) => {
+    if (sortType === '가나다순') {
+      return [...clubs].sort((a, b) => a.name.localeCompare(b.name));
+    } else {
+      return [...clubs].sort((a, b) => b.views - a.views);
+    }
+  };
+
+  const filteredClubs = getSortedClubs(clubs).filter(
+    club => clubCategoryMap[club.clubType] === selectedCategory,
+  );
 
   return (
     <SafeAreaView style={backgroundStyle}>
@@ -86,9 +113,50 @@ const AssociationScreen: React.FC<AssociationScreenProps> = ({navigation}) => {
           onPress={() => navigation.goBack()}>
           <Text style={[styles.backButtonText, {color: textColor}]}>뒤로</Text>
         </TouchableOpacity>
-        <Text style={[styles.headerTitle, {color: textColor}]}>자치단체</Text>
+        <Text style={[styles.headerTitle, {color: textColor}]}>동아리</Text>
         <View style={styles.placeholderButton} />
       </View>
+
+      <ScrollView
+        horizontal
+        showsHorizontalScrollIndicator={false}
+        style={styles.typeNav}>
+        {clubCategories.map(category => (
+          <View key={category} style={styles.typeTabWrapper}>
+            <View style={styles.typeTabInner}>
+              <View style={styles.textWithUnderline}>
+                <Text
+                  style={[
+                    styles.typeTab,
+                    {color: isDarkMode ? '#888' : '#999'},
+                    selectedCategory === category && [
+                      styles.selectedTypeText,
+                      {color: textColor},
+                    ],
+                  ]}
+                  onPress={() => setSelectedCategory(category)}
+                  onLayout={e => {
+                    const width = e.nativeEvent.layout.width;
+                    setTextWidths(prev => ({...prev, [category]: width}));
+                  }}>
+                  {category}
+                </Text>
+                {selectedCategory === category && (
+                  <View
+                    style={[
+                      styles.underline,
+                      {
+                        width: (textWidths[category] || 0) + 8,
+                        backgroundColor: textColor,
+                      },
+                    ]}
+                  />
+                )}
+              </View>
+            </View>
+          </View>
+        ))}
+      </ScrollView>
 
       <View style={styles.sortButtons}>
         {sortTypes.map(type => (
@@ -128,51 +196,51 @@ const AssociationScreen: React.FC<AssociationScreenProps> = ({navigation}) => {
             ]}>
             <Text style={[styles.tempText, {color: textColor}]}>로딩중...</Text>
           </View>
-        ) : associations.length === 0 ? (
+        ) : filteredClubs.length === 0 ? (
           <View
             style={[
               styles.contentContainer,
               {backgroundColor: isDarkMode ? '#1A1A1A' : '#fff'},
             ]}>
             <Text style={[styles.tempText, {color: textColor}]}>
-              등록된 자치단체가 없습니다.
+              등록된 동아리가 없습니다.
             </Text>
           </View>
         ) : (
-          getSortedAssociations(associations).map(association => (
+          filteredClubs.map(club => (
             <TouchableOpacity
-              key={association.uuid}
+              key={club.uuid}
               style={[
-                styles.associationCard,
+                styles.clubCard,
                 {backgroundColor: isDarkMode ? '#1A1A1A' : '#fff'},
               ]}
               onPress={() =>
-                navigation.navigate('AssociationDetail', {
-                  associationId: association.uuid,
-                  associationName: association.name,
+                navigation.navigate('ClubDetail', {
+                  clubId: club.uuid,
+                  clubName: club.name,
                 })
               }>
               <View style={styles.cardContent}>
                 <View style={styles.imageContainer}>
                   <Image
-                    source={{uri: association.image_url}}
-                    style={styles.associationImage}
+                    source={{uri: club.image_url}}
+                    style={styles.clubImage}
                     resizeMode="contain"
                   />
                 </View>
-                <View style={styles.associationInfo}>
+                <View style={styles.clubInfo}>
                   <Text
-                    style={[styles.associationName, {color: textColor}]}
+                    style={[styles.clubName, {color: textColor}]}
                     numberOfLines={1}>
-                    {association.name}
+                    {club.name}
                   </Text>
                   <Text
                     style={[
-                      styles.associationDescription,
+                      styles.clubDescription,
                       {color: isDarkMode ? '#888' : '#666'},
                     ]}
                     numberOfLines={2}>
-                    {association.content}
+                    {club.short_desc}
                   </Text>
                 </View>
               </View>
@@ -203,6 +271,56 @@ const styles = StyleSheet.create({
     flex: 1,
     textAlign: 'center',
   },
+  typeNav: {
+    paddingLeft: 16,
+    paddingVertical: 8,
+    minHeight: 52,
+  },
+  typeTabWrapper: {
+    paddingVertical: 8,
+    marginRight: 24,
+    height: 36,
+    justifyContent: 'center',
+  },
+  typeTabInner: {
+    alignItems: 'center',
+    justifyContent: 'center',
+    minHeight: 24,
+  },
+  textWithUnderline: {
+    alignItems: 'center',
+    paddingBottom: 4,
+  },
+  typeTab: {
+    fontSize: 16,
+    color: '#999',
+    paddingHorizontal: 4,
+  },
+  selectedTypeText: {
+    color: '#000',
+    fontWeight: 'bold',
+  },
+  underline: {
+    marginTop: 4,
+    height: 2,
+    borderRadius: 1,
+  },
+  contentContainer: {
+    padding: 16,
+    borderRadius: 12,
+    shadowColor: '#000',
+    shadowOffset: {
+      width: 0,
+      height: 2,
+    },
+    shadowOpacity: 0.1,
+    shadowRadius: 4,
+    elevation: 6,
+  },
+  tempText: {
+    fontSize: 16,
+    textAlign: 'center',
+  },
   backButton: {
     padding: 8,
   },
@@ -224,21 +342,7 @@ const styles = StyleSheet.create({
     borderWidth: 1,
     marginRight: 10,
   },
-  activeSort: {
-    borderColor: '#000',
-  },
-  sortButtonText: {
-    fontSize: 14,
-  },
-  contentContainer: {
-    padding: 16,
-    borderRadius: 12,
-  },
-  tempText: {
-    fontSize: 16,
-    textAlign: 'center',
-  },
-  associationCard: {
+  clubCard: {
     marginBottom: 12,
     borderRadius: 12,
     shadowColor: '#000',
@@ -264,24 +368,30 @@ const styles = StyleSheet.create({
     justifyContent: 'center',
     alignItems: 'center',
   },
-  associationImage: {
-    padding: 10,
+  clubImage: {
+    padding: 5,
     width: '100%',
     height: '100%',
   },
-  associationInfo: {
+  clubInfo: {
     flex: 1,
     marginLeft: 16,
   },
-  associationName: {
+  clubName: {
     fontSize: 16,
     fontWeight: '600',
     marginBottom: 4,
   },
-  associationDescription: {
+  clubDescription: {
     fontSize: 14,
     lineHeight: 20,
   },
+  activeSort: {
+    borderColor: '#000',
+  },
+  sortButtonText: {
+    fontSize: 14,
+  },
 });
 
-export default AssociationScreen;
+export default ClubScreen;
